refactor container and deployments. scope logic appropriately and remove unused code.
