+++ crates/utils/src/text.rs
@@ -22,3 +22,39 @@ pub fn short_uuid(u: &Uuid) -> String {
     let full = u.simple().to_string();
     full.chars().take(4).collect() // grab the first 4 chars
 }
+
+pub fn truncate_to_char_boundary(content: &str, max_len: usize) -> &str {
+    if content.len() <= max_len {
+        return content;
+    }
+
+    let cutoff = content
+        .char_indices()
+        .map(|(idx, _)| idx)
+        .chain(std::iter::once(content.len()))
+        .take_while(|&idx| idx <= max_len)
+        .last()
+        .unwrap_or(0);
+
+    debug_assert!(content.is_char_boundary(cutoff));
+    &content[..cutoff]
+}
+
+#[cfg(test)]
+mod tests {
+
+    #[test]
+    fn test_truncate_to_char_boundary() {
+        use super::truncate_to_char_boundary;
+
+        let input = "a".repeat(10);
+        assert_eq!(truncate_to_char_boundary(&input, 7), "a".repeat(7));
+
+        let input = "hello world";
+        assert_eq!(truncate_to_char_boundary(input, input.len()), input);
+
+        let input = "🔥🔥🔥"; // each fire emoji is 4 bytes
+        assert_eq!(truncate_to_char_boundary(input, 5), "🔥");
+        assert_eq!(truncate_to_char_boundary(input, 3), "");
+    }
+}
